+++ src/onegov.activity/onegov/activity/collections/__init__.py
@@ -2,7 +2,7 @@ from onegov.activity.collections.activity import ActivityFilter
 from onegov.activity.collections.activity import ActivityCollection
 from onegov.activity.collections.attendee import AttendeeCollection
 from onegov.activity.collections.booking import BookingCollection
-from onegov.activity.collections.invoice_item import InvoiceItemCollection
+from onegov.activity.collections.invoice import InvoiceCollection
 from onegov.activity.collections.occasion import OccasionCollection
 from onegov.activity.collections.period import PeriodCollection
 from onegov.activity.collections.publication_request import \
@@ -13,7 +13,7 @@ __all__ = [
     'ActivityFilter',
     'AttendeeCollection',
     'BookingCollection',
-    'InvoiceItemCollection',
+    'InvoiceCollection',
     'OccasionCollection',
     'PeriodCollection',
     'PublicationRequestCollection'