+++ src/onegov.activity/onegov/activity/models/invoice_item.py
@@ -1,18 +1,22 @@
-from onegov.activity import utils
 from onegov.core.orm import Base
 from onegov.core.orm.mixins import TimestampMixin
 from onegov.core.orm.types import UUID
 from onegov.pay import PayableManyTimes
-from onegov.user import User
 from sqlalchemy import Boolean
 from sqlalchemy import Column
 from sqlalchemy import ForeignKey
 from sqlalchemy import Numeric
 from sqlalchemy import Text
 from sqlalchemy.ext.hybrid import hybrid_property
-from sqlalchemy.orm import validates, relationship
-from sqlalchemy_utils import observes
+from sqlalchemy.orm import validates
 from uuid import uuid4
+
+
+# total digits
+PRECISION = 8
+
+# digits after the point
+SCALE = 2
 
 
 class InvoiceItem(Base, TimestampMixin, PayableManyTimes):
@@ -23,15 +27,8 @@ class InvoiceItem(Base, TimestampMixin, PayableManyTimes):
     #: the public id of the invoice item
     id = Column(UUID, primary_key=True, default=uuid4)
 
-    #: the user who is charged
-    username = Column(Text, ForeignKey('users.username'), nullable=False)
-    user = relationship(User, backref="invoice_items")
-
-    #: the invoice group (all items with the same text make one invoice)
-    invoice = Column(Text, nullable=False)
-
-    #: the code of the invoice used to identify the invoice through e-banking
-    code = Column(Text, nullable=False)
+    #: the invoice this item belongs to
+    invoice_id = Column(UUID, ForeignKey('invoices.id'))
 
     #: the item group (all items with the same text are visually grouped)
     group = Column(Text, nullable=False)
@@ -52,51 +49,21 @@ class InvoiceItem(Base, TimestampMixin, PayableManyTimes):
     source = Column(Text, nullable=True)
 
     #: the unit to pay..
-    unit = Column(Numeric(precision=8, scale=2), nullable=True)
+    unit = Column(Numeric(precision=PRECISION, scale=SCALE), nullable=True)
 
     #: ..multiplied by the quantity..
-    quantity = Column(Numeric(precision=8, scale=2), nullable=True)
+    quantity = Column(Numeric(precision=PRECISION, scale=SCALE), nullable=True)
 
     #: ..together form the amount
     @hybrid_property
     def amount(self):
-        return self.unit * self.quantity
+        return round(self.unit, SCALE) * round(self.quantity, SCALE)
 
-    @observes('invoice', 'username')
-    def invoice_username_observer(self, invoice, username):
-        self.code = utils.as_invoice_code(invoice, username)
+    @amount.expression
+    def amount(cls):
+        return cls.unit * cls.quantity
 
     @validates('source')
     def validate_source(self, key, value):
         assert value in (None, 'xml', 'stripe_connect')
         return value
-
-    @property
-    def display_code(self):
-        """ The item's code, formatted in a human readable format. """
-        return utils.format_invoice_code(self.code)
-
-    @property
-    def esr_code(self):
-        """ The item's code, formatted as ESR. """
-        return utils.encode_invoice_code(self.code)
-
-    @property
-    def display_esr_code(self):
-        """ The item's ESR formatted in a human readable format. """
-        return utils.format_esr_reference(self.esr_code)
-
-    @property
-    def discourage_changes(self):
-        return self.source == 'xml'
-
-    @property
-    def disable_changes(self):
-        if not self.source:
-            return False
-
-        if self.source == 'xml':
-            return False
-
-        states = {p.state for p in self.payments}
-        return 'open' in states or 'paid' in states