+++ src/onegov.activity/onegov/activity/tests/test_iso20022.py
@@ -1,6 +1,6 @@
 from datetime import date
 from decimal import Decimal
-from onegov.activity.collections import InvoiceItemCollection
+from onegov.activity.collections import InvoiceCollection
 from onegov.activity.iso20022 import extract_transactions, extract_code
 from onegov.activity.iso20022 import match_iso_20022_to_usernames
 from onegov.activity.utils import generate_xml, encode_invoice_code
@@ -114,53 +114,62 @@ def test_extract_code():
     assert extract_code('Code: q-70171-292fa') == 'q70171292fa'
 
 
-def test_invoice_matching(session, owner, member):
-    items = InvoiceItemCollection(session)
+def test_invoice_matching(session, owner, member,
+                          prebooking_period, inactive_period):
 
-    i1 = items.add(owner, '2017', 'Aaron', 'Billard', 250, 1)
-    i2 = items.add(owner, '2017', 'Baron', 'Pokemon', 250, 1)
+    p1 = prebooking_period
+    p2 = inactive_period
 
-    assert i1.code == i2.code
+    invoices = InvoiceCollection(session)
 
-    i3 = items.add(member, '2017', 'Connie', 'Swimming', 250, 1)
-    i4 = items.add(member, '2018', 'Donnie', 'Football', 250, 1)
+    own_i1 = invoices.add(user_id=owner.id, period_id=p1.id)
+    mem_i1 = invoices.add(user_id=member.id, period_id=p1.id)
+    mem_i2 = invoices.add(user_id=member.id, period_id=p2.id)
 
-    assert i3.code != i4.code
+    i1 = own_i1.add('Aaron', 'Billard', 250, 1)
+    i2 = own_i1.add('Baron', 'Pokemon', 250, 1)
+
+    assert i1.invoice.code == i2.invoice.code
+
+    i3 = mem_i1.add('Connie', 'Swimming', 250, 1)
+    i4 = mem_i2.add('Donnie', 'Football', 250, 1)
+
+    assert i3.invoice.code != i4.invoice.code
 
     # match against a perfect match and one with a wrong amount
     xml = generate_xml([
-        dict(amount='500.00 CHF', note=i1.code),
-        dict(amount='500.00 CHF', note=i3.code)
+        dict(amount='500.00 CHF', note=i1.invoice.code),
+        dict(amount='500.00 CHF', note=i3.invoice.code)
     ])
 
-    transactions = list(match_iso_20022_to_usernames(xml, items, '2017'))
+    transactions = list(match_iso_20022_to_usernames(xml, session, p1.id))
 
     assert len(transactions) == 2
     assert transactions[0].amount == Decimal(500.00)
     assert transactions[0].username == owner.username
-    assert transactions[0].note == i1.code == i2.code
+    assert transactions[0].note == i1.invoice.code == i2.invoice.code
     assert transactions[0].confidence == 1.0
     assert transactions[1].amount == Decimal(500.00)
     assert transactions[1].username == member.username
-    assert transactions[1].note == i3.code
+    assert transactions[1].note == i3.invoice.code
     assert transactions[1].confidence == 0.5
 
     # debit transactions are simply ignored
     xml = generate_xml([
-        dict(amount='-500.00 CHF', note=i1.code),
-        dict(amount='-500.00 CHF', note=i3.code)
+        dict(amount='-500.00 CHF', note=i1.invoice.code),
+        dict(amount='-500.00 CHF', note=i3.invoice.code)
     ])
 
-    transactions = list(match_iso_20022_to_usernames(xml, items, '2017'))
+    transactions = list(match_iso_20022_to_usernames(xml, session, p1.id))
 
     assert len(transactions) == 0
 
     # match against a code with extra information
     xml = generate_xml([
-        dict(amount='500.00 CHF', note='Code: ' + i1.code),
+        dict(amount='500.00 CHF', note='Code: ' + i1.invoice.code),
     ])
 
-    transactions = list(match_iso_20022_to_usernames(xml, items, '2017'))
+    transactions = list(match_iso_20022_to_usernames(xml, session, p1.id))
 
     assert len(transactions) == 1
     assert transactions[0].username == owner.username
@@ -168,12 +177,12 @@ def test_invoice_matching(session, owner, member):
 
     # match against transactions from a different invoice
     xml = generate_xml([
-        dict(amount='500.00 CHF', note=i1.code),
-        dict(amount='250.00 CHF', note=i3.code),
-        dict(amount='250.00 CHF', note=i4.code),
+        dict(amount='500.00 CHF', note=i1.invoice.code),
+        dict(amount='250.00 CHF', note=i3.invoice.code),
+        dict(amount='250.00 CHF', note=i4.invoice.code),
     ])
 
-    transactions = list(match_iso_20022_to_usernames(xml, items, '2018'))
+    transactions = list(match_iso_20022_to_usernames(xml, session, p2.id))
 
     assert len(transactions) == 3
     assert transactions[0].amount == Decimal(500.00)
@@ -188,10 +197,10 @@ def test_invoice_matching(session, owner, member):
 
     # match against the wrong currency
     xml = generate_xml([
-        dict(amount='250.00 EUR', note=i4.code),
+        dict(amount='250.00 EUR', note=i4.invoice.code),
     ])
 
-    transactions = list(match_iso_20022_to_usernames(xml, items, '2018'))
+    transactions = list(match_iso_20022_to_usernames(xml, session, p2.id))
 
     assert len(transactions) == 1
     assert transactions[0].username is None
@@ -202,7 +211,7 @@ def test_invoice_matching(session, owner, member):
         dict(amount='250.00 CHF', note='asdf'),
     ])
 
-    transactions = list(match_iso_20022_to_usernames(xml, items, '2018'))
+    transactions = list(match_iso_20022_to_usernames(xml, session, p2.id))
     assert len(transactions) == 1
     assert transactions[0].username == member.username
     assert transactions[0].confidence == 0.5
@@ -212,18 +221,18 @@ def test_invoice_matching(session, owner, member):
         dict(amount='123.00 CHF', note='asdf'),
     ])
 
-    transactions = list(match_iso_20022_to_usernames(xml, items, '2018'))
+    transactions = list(match_iso_20022_to_usernames(xml, session, p2.id))
     assert len(transactions) == 1
     assert transactions[0].username is None
     assert transactions[0].confidence == 0
 
     # match against duplicate bookings
     xml = generate_xml([
-        dict(amount='250.00 CHF', note=i4.code),
-        dict(amount='250.00 CHF', note=i4.code),
+        dict(amount='250.00 CHF', note=i4.invoice.code),
+        dict(amount='250.00 CHF', note=i4.invoice.code),
     ])
 
-    transactions = list(match_iso_20022_to_usernames(xml, items, '2018'))
+    transactions = list(match_iso_20022_to_usernames(xml, session, p2.id))
     assert len(transactions) == 2
     assert transactions[0].username == member.username
     assert transactions[0].confidence == 1.0
@@ -234,11 +243,11 @@ def test_invoice_matching(session, owner, member):
 
     # match against split bookings
     xml = generate_xml([
-        dict(amount='125.00 CHF', note=i1.code),
-        dict(amount='125.00 CHF', note=i1.code),
+        dict(amount='125.00 CHF', note=i1.invoice.code),
+        dict(amount='125.00 CHF', note=i1.invoice.code),
     ])
 
-    transactions = list(match_iso_20022_to_usernames(xml, items, '2018'))
+    transactions = list(match_iso_20022_to_usernames(xml, session, p2.id))
     assert len(transactions) == 2
     assert transactions[0].username is None
     assert transactions[0].confidence == 0
@@ -249,10 +258,12 @@ def test_invoice_matching(session, owner, member):
 
     # match against reference numbers
     xml = generate_xml([
-        dict(amount='500.00 CHF', reference=encode_invoice_code(i1.code)),
+        dict(amount='500.00 CHF', reference=encode_invoice_code(
+            i1.invoice.code
+        )),
     ])
 
-    transactions = list(match_iso_20022_to_usernames(xml, items, '2017'))
+    transactions = list(match_iso_20022_to_usernames(xml, session, p1.id))
 
     assert len(transactions) == 1
     assert transactions[0].username == owner.username
@@ -260,11 +271,15 @@ def test_invoice_matching(session, owner, member):
 
     # match against a mix of reference numbers and codes
     xml = generate_xml([
-        dict(amount='500.00 CHF', reference=encode_invoice_code(i1.code)),
-        dict(amount='250.00 CHF', note=i3.code),
+        dict(
+            amount='500.00 CHF',
+            reference=encode_invoice_code(i1.invoice.code)),
+        dict(
+            amount='250.00 CHF',
+            note=i3.invoice.code),
     ])
 
-    transactions = list(match_iso_20022_to_usernames(xml, items, '2017'))
+    transactions = list(match_iso_20022_to_usernames(xml, session, p1.id))
 
     assert len(transactions) == 2
     assert transactions[0].username == owner.username
@@ -275,10 +290,10 @@ def test_invoice_matching(session, owner, member):
     # ignore invalid esr references
     xml = generate_xml([
         dict(amount='500.00 CHF', reference='booyah'),
-        dict(amount='250.00 CHF', note=i3.code),
+        dict(amount='250.00 CHF', note=i3.invoice.code),
     ])
 
-    transactions = list(match_iso_20022_to_usernames(xml, items, '2017'))
+    transactions = list(match_iso_20022_to_usernames(xml, session, p1.id))
 
     assert len(transactions) == 2
     assert transactions[0].username == owner.username
@@ -291,12 +306,12 @@ def test_invoice_matching(session, owner, member):
     xml = generate_xml([
         dict(
             amount='500.00 CHF',
-            reference=encode_invoice_code(i1.code),
-            note=i1.code
+            reference=encode_invoice_code(i1.invoice.code),
+            note=i1.invoice.code
         ),
     ])
 
-    transactions = list(match_iso_20022_to_usernames(xml, items, '2017'))
+    transactions = list(match_iso_20022_to_usernames(xml, session, p1.id))
 
     assert len(transactions) == 1
     assert transactions[0].username == owner.username
@@ -306,11 +321,11 @@ def test_invoice_matching(session, owner, member):
         dict(
             amount='500.00 CHF',
             reference='000127131108141601011502061',
-            note=i1.code
+            note=i1.invoice.code
         ),
     ])
 
-    transactions = list(match_iso_20022_to_usernames(xml, items, '2017'))
+    transactions = list(match_iso_20022_to_usernames(xml, session, p1.id))
 
     assert len(transactions) == 1
     assert transactions[0].username == owner.username
@@ -319,12 +334,12 @@ def test_invoice_matching(session, owner, member):
     xml = generate_xml([
         dict(
             amount='500.00 CHF',
-            reference=encode_invoice_code(i1.code),
+            reference=encode_invoice_code(i1.invoice.code),
             note='qeb3afd0e43'
         ),
     ])
 
-    transactions = list(match_iso_20022_to_usernames(xml, items, '2017'))
+    transactions = list(match_iso_20022_to_usernames(xml, session, p1.id))
 
     assert len(transactions) == 1
     assert transactions[0].username == owner.username
@@ -337,11 +352,11 @@ def test_invoice_matching(session, owner, member):
     i1.source = 'xml'
 
     xml = generate_xml([
-        dict(amount='250 CHF', note=i1.code, tid='foobar'),
-        dict(amount='250 CHF', note=i1.code, tid=None),
+        dict(amount='250 CHF', note=i1.invoice.code, tid='foobar'),
+        dict(amount='250 CHF', note=i1.invoice.code, tid=None),
     ])
 
-    transactions = list(match_iso_20022_to_usernames(xml, items, '2017'))
+    transactions = list(match_iso_20022_to_usernames(xml, session, p1.id))
     assert len(transactions) == 2
 
     assert transactions[0].tid == 'foobar'